+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Fixed ability to handle ChemKED files with uncertainty for various properties.
+- Updated handling of RCM volume history and compression time to PyKED v0.4.1
 
 ### Changed
 - Removed interpolation from end of integration (was never really necessary).